refactor

+++ src/states/gamestate.c
@@ -61,7 +61,7 @@ bool finishGameState(void)
 // Logic Functions
 //----------------------------------------------------------------------------------
 #define PIPE_SPAWN_PADDING (20)
-#define MIN_PIPE_Y (0 - pipe->pipeTop->height)
+#define MIN_PIPE_Y (0 - (int) pipe->pipeTop->height)
 #define MAX_PIPE_Y (GetScreenHeight() - (int) pipe->pipeTop->height - (int) pipe->pipeGap)
 
 static void spawnPipe(const float deltaTime)
